Bitcoin Blockchain Client Software Bitcoin Core Update
Bitcoin Blockchan and the Bitcoin Network Interface the Client Software is going through a series update under the supervised initiation by Satoshi Nakamoto.  Satoshi Nakamoto has carried out a complex research over the last 10 years and has uncovered the quantum logic and developed  new computational ptocol for the Bitcoin Core updating.  Which will take effect by end of the year 2023.   In the beginning of 2024 the Bitcoin Blockchain will improved with new quantum  network interface and Proof-of-Authentication instead of proof-of-work and proof-of-stake. 

Satoshi Nakamoto has developed  few more innovative invention that will be embedded in the Bitcoin Blockchain for complete security.  
The improvement will be carried by Satoshi Nakamoto for the mass adoption of Bitcoin and Blockchain.  His plan is to find solution to the Bitcoin battle between maximalist and minimalist.  What we need is a new approach to greater social good instead of self gratification.  The game changing theory is based on the time machine and the time machin is not control by any syndicated group of previleged Elites.  We are here to establish another group of new elite like those of Central banking families.  We need to focus on complete financial and economic liberation for mankind.